// src/app/api/models/api-response-notificacion-configracion-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { NotificacionConfigracionResponse } from '../models/notificacion-configracion-response';
export interface ApiResponseNotificacionConfigracionResponse {
  data?: NotificacionConfigracionResponse;
  message?: string;
  success?: boolean;
}